+++ app/src/main/java/eu/faircode/email/DB.java
@@ -68,7 +68,7 @@ import javax.mail.internet.InternetAddress;
 // https://developer.android.com/topic/libraries/architecture/room.html
 
 @Database(
-        version = 274,
+        version = 275,
         entities = {
                 EntityIdentity.class,
                 EntityAccount.class,
@@ -2780,6 +2780,12 @@ public abstract class DB extends RoomDatabase {
                         logMigration(startVersion, endVersion);
                         db.execSQL("ALTER TABLE `answer` ADD COLUMN `label` TEXT");
                         db.execSQL("CREATE UNIQUE INDEX IF NOT EXISTS `index_answer_label` ON `answer` (`label`)");
+                    }
+                }).addMigrations(new Migration(274, 275) {
+                    @Override
+                    public void migrate(@NonNull SupportSQLiteDatabase db) {
+                        logMigration(startVersion, endVersion);
+                        db.execSQL("ALTER TABLE `contact` ADD COLUMN `folder` INTEGER");
                     }
                 }).addMigrations(new Migration(998, 999) {
                     @Override