+++ CHANGELOG.md
@@ -6,6 +6,12 @@ For support you can use [the contact form](https://contact.faircode.eu/?product=
 
 ### [Rahonavis](https://en.wikipedia.org/wiki/Rahonavis)
 
+### Next version
+
+* Added option to configure number of widget subject lines
+* Small improvements and minor bug fixes
+* Updated [translations](https://crowdin.com/project/open-source-email)
+
 ### 1.2054 - 2023-03-14
 
 * Added rule action option to delete keywords