+++ CHANGELOG.md
@@ -9,6 +9,7 @@ For support you can use [the contact form](https://contact.faircode.eu/?product=
 ### Next version
 
 * Added option to configure number of widget subject lines
+* Added remembering last used folder for contacts
 * Small improvements and minor bug fixes
 * Updated [translations](https://crowdin.com/project/open-source-email)